backend/requirements.txt
Django==4.0.10
djangorestframework==3.14.0
djangorestframework_simplejwt== 5.3.1
PyJWT==2.8.0
PyJWT==2.8.0
djongo==1.3.6
pymongo==3.12.2
